Ignore properties of b not in a.
Previously, the implementation of d3.interpolate{Array,Object} were not
consistent with their documentation, which described how elements or properties
in a but not in b were ignored. This commit changes the implementation to ignore
those properties as intended, such that the final result of the interpolation is
more consistent with the specified end value b.

Fixes #19. Fixes #20.