interpolateString: clarify docs
The example paragraph for `interpolateString` was worded ambiguously ("The remaining static parts of the string", without specifying which string is being talked about); so unless one had already grasped how the function is supposed to work from the abstract description in the paragraphs above, the interpolation result would be surprising and unclear. The documentation thus fails "bottom-up" learners (i.e. those who learn best by example).

This change is a proposed clarification of the example paragraph by explicitly mentioning the *b* string.